requirements.txt
opencc-python-reimplemented
ebooklib
beautifulsoup4
customtkinter
tkinterdnd2-universal